refactor: fix and remove non utilized imports
