+++ navigation/CreateNewWallet.tsx
@@ -6,7 +6,7 @@ import Text from '../components/Text';
 import Container from '../components/Container';
 import AnimatedStep from '../components/AnimatedStep';
 import {KeyboardAwareScrollView} from 'react-native-keyboard-aware-scroll-view';
-import useWallet from '../hooks/useWallet';
+import useWallet from '../src/hooks/useWallet';
 import LoadingModalContent from '../components/Modals/LoadingModalContent';
 import {NetworkTypes} from '../constants/Type';
 import OptionCard from '../components/OptionCard';
@@ -14,8 +14,8 @@ import Mnemonic from '../components/Mnemonic';
 import {layoutStyles} from '../utils/layout';
 import TopNavigationComponent from '../components/TopNavigation';
 import {scale, verticalScale} from 'react-native-size-matters';
-import useSecurity from '../hooks/useSecurity';
-import {useModal} from '../hooks/useModal';
+import useSecurity from '../src/hooks/useSecurity';
+import {useModal} from '../src/hooks/useModal';
 import {errorTextParser, promptErrorToaster} from '../utils/errors';
 import ErrorModalContent from '../components/Modals/ErrorModalContent';
 import SkipMnemonicConfirmationModalContent from '../components/Modals/SkipMnemonicConfirmationModalContent';
