+++ navigation/OpenWallet.tsx
@@ -3,14 +3,14 @@ import {StyleSheet, View} from 'react-native';
 import {useNavigation} from '@react-navigation/native';
 
 import Container from '../components/Container';
-import useWallet from '../hooks/useWallet';
+import useWallet from '../src/hooks/useWallet';
 import LoadingModalContent from '../components/Modals/LoadingModalContent';
 import OptionCard from '../components/OptionCard';
 import {KeyboardAwareScrollView} from 'react-native-keyboard-aware-scroll-view';
 import {layoutStyles} from '../utils/layout';
 import TopNavigationComponent from '../components/TopNavigation';
-import useSecurity from '../hooks/useSecurity';
-import {useModal} from '../hooks/useModal';
+import useSecurity from '../src/hooks/useSecurity';
+import {useModal} from '../src/hooks/useModal';
 import {errorTextParser, promptErrorToaster} from '../utils/errors';
 import ErrorModalContent from '../components/Modals/ErrorModalContent';
